feat(config): autoload "extension bins"
The extension bins are directories in the nushell config folder that should be
automatically sourced or executed upon shell startup. The advantage of doing it
this way is component modularity, since we can identify the purpose of each
sourced script via their parent directories and moreover we have some extra
granularity in terms of the order or parameterization of the loading process.