fix(autoload): remove a few non-existent autoloads
In the previous repo, there were a few things that were not being used, and have
become obsolete. This includes things like a `weechat` module, which was neither
functional nor a very useful idea to begin with. Hence, we need to adjust the
autoload directives so that we don't error at shell startup.