feat(nupm): add `nupm` submodule via nushell
This is somewhat convoluted. Essentially, I have a hook set up to use the nupm
module as an overlay, but the module source comes directly via github and the
nushell repository for `nupm`. But, `jj` does not support submodules directly.
Therefore, in order to include this as a submodule, we need to use `git`
directly.

If this were a colocated `git` repo (which I maybe should have considered), I
would have no problem just doing the git commands directly. However, this is not
colocated, and as such in order to use `git` commands we need to set up some
environment variables that point `git` in the correct direction of this
"worktree".

This is handled via the .envrc file present here. However, we don't want to
further commit any of the `nupm` source files to our own repo, as they are
handled elsewhere. Hence the inclusion of `nupm` in the gitignore file that
`jj` uses. Recall the use of `jj file untrack` if there is confusion about why
the nupm submodule is still being tracked despite being specified as ignorable.
The `.gitmodules` file is simply a standard definition of the `nupm` submodule.