feat(alias): add alias that toggles the lock state of kde panels
I don't remember exactly when, but at some point my KDE panel was no
longer able to be edited; this is to say that the user could not select
any options to unlock the panels for editing. Some googling led me to
this command which I have no added as a permanent alias.